% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeConditions.R
\name{treeConditions}
\alias{treeConditions}
\title{Tree Conditions}
\usage{
treeConditions(in.tree, split.digits = 11)
}
\arguments{
\item{in.tree}{Trees from a fitted XGBoost model}

\item{split.digits}{Significant digits for rounding split values}
}
\value{
Returns a table of tree conditions.
}
\description{
Extract all the conditional representations of a model.
}
\details{
Search the trees of an XGBoost model and convert them into feature conditions and values.
Combine values across the same feature conditions.
}
