% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeFormula.R
\name{treeFormula}
\alias{treeFormula}
\title{Tree Formula}
\usage{
treeFormula(in.conditions)
}
\arguments{
\item{in.conditions}{input table of conditions}
}
\value{
A new data.table with formula columns attached to the input.
\itemize{
  \item \code{formula}: overall formula in the form of a string
  \item \code{formula_i}: string formula for the i-th feature
}
}
\description{
Attach formula columns from an input condition table.
}
