% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeInteractions.R
\name{treeInteractions}
\alias{treeInteractions}
\title{Tree Interactions}
\usage{
treeInteractions(in.tree, in.depth)
}
\arguments{
\item{in.tree}{Trees from a fitted XGBoost model}

\item{in.depth}{Maximum depth of the input model}
}
\value{
Returns a list of vectors where each item of the list contains the feature names of one interaction.
}
\description{
Extract interactions fitted in an XGBoost model.
}
\details{
Search the trees of an XGBoost model to identify fitted interactions.
For tree based models, interactions are modelled by features sharing the same tree.
Pads out lower order interactions that are contained within a higher order interaction.
}
