% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeRanks.R
\name{treeRanks}
\alias{treeRanks}
\title{Tree Ranks}
\usage{
treeRanks(in.conditions, in.data, in.weights = NULL)
}
\arguments{
\item{in.conditions}{table of normalised tree conditions}

\item{in.data}{input dataset to calculate ranks}

\item{in.weights}{vector of weights corresponding to each observation of \code{in.data}, defaults to a vector of 1s}
}
\value{
A table of parameters and their associated importance score.
}
\description{
Rank the importance of each parameter on an input dataset
}
