% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeScores.R
\name{treeScores}
\alias{treeScores}
\title{Tree Score}
\usage{
treeScores(in.conditions, in.data)
}
\arguments{
\item{in.conditions}{table containing a formula column defining}

\item{in.data}{input data to score, in the form of a \code{data.table}}
}
\description{
Score using formulas and values from an input conditions table
}
\details{
Calculate scores from a conditions table.  Useful for reconciling with the original XGBoost model.
}
